LEARN use of global variables exports - and example of encapsulation
